kfs: memory leaks bug fixes
vmm: invalidate page with defer to invalidate in error paths
vheap: alignment to PAGE_SIZE for more efficient free list for big allocations
delete and release mapping support now switch to vas to use delete in error path
    of clone / fork and prepareBinary
clone / fork: proper delete for MM in case of error
prepareBinary: release allocated mappings in case od error

+++ src/kernel/mm/heap.zig
@@ -26,7 +26,7 @@ pub const FreeList = struct {
     head: ?*FreeListNode,
     vmm: *vmm,
     pmm: *pmm,
-    alignement: u8 = 16,
+    alignement: usize = 16,
     mtx: krn.Spinlock,
 
     pub fn init(
@@ -34,7 +34,7 @@ pub const FreeList = struct {
         virt_mm: *vmm,
         start_addr: usize,
         end_addr: usize,
-        alignement: u8,
+        alignement: usize,
     ) FreeList {
         return FreeList{
             .start_addr = start_addr,
